+++ data_script/data_preparation.py
@@ -1,0 +1,3 @@
+from data_script.read_gsheet import get_specific_range
+
+get_specific_range('unpreparing_stady_data')
